+++ CHANGELOG.md
@@ -38,6 +38,7 @@ See <https://commonmark.thephpleague.com/2.0/upgrading/> for detailed informatio
      - `containsKey()`
      - `replaceWith()`
      - `removeGaps()`
+   - Removed the `ListBlock::TYPE_UNORDERED` constant
  - Removed the unused `Delimiter::setCanClose()` method
 
 [unreleased]: https://github.com/thephpleague/commonmark/compare/1.4...master